Rewire ProAi hierarchy
