+++ game-core/src/main/java/games/strategy/triplea/ai/pro/logging/ProLogger.java
@@ -1,5 +1,6 @@
 package games.strategy.triplea.ai.pro.logging;
 
+import games.strategy.triplea.ai.pro.AbstractProAi;
 import java.util.logging.Level;
 import javax.annotation.Nullable;
 
@@ -30,17 +31,17 @@ public final class ProLogger {
   /**
    * Some notes on using the Pro AI logger: First, to make the logs easily readable even when there
    * are hundreds of lines, I want every considerable step down in the call stack to mean more log
-   * message indentation. For example, the base logs in the {@link
-   * games.strategy.triplea.ai.pro.ProAi} class have no indentation before them, but the base logs
-   * in the {@link games.strategy.triplea.ai.pro.ProCombatMoveAi} class will have two spaces
-   * inserted at the start, and the level below that, four spaces. In this way, when you're reading
-   * the log, you can skip over unimportant areas with speed because of the indentation. Second, I
-   * generally want the Fine logs to be messages that run less than 10 times each round, including
-   * almost all messages in the Pro AI class. Finest for messages showing details within a method
-   * that, for example, returns a value. (So, for example, the NCM_Task method IsTaskWorthwhile()
-   * would primarily use finest, as it just returns a boolean, and the logs within it are just for
-   * details) Finer for just about everything else. (There's also the SERVER, INFO, etc. levels)
-   * Just keep these things in mind while adding new logging code.
+   * message indentation. For example, the base logs in the {@link AbstractProAi} class have no
+   * indentation before them, but the base logs in the {@link
+   * games.strategy.triplea.ai.pro.ProCombatMoveAi} class will have two spaces inserted at the
+   * start, and the level below that, four spaces. In this way, when you're reading the log, you can
+   * skip over unimportant areas with speed because of the indentation. Second, I generally want the
+   * Fine logs to be messages that run less than 10 times each round, including almost all messages
+   * in the Pro AI class. Finest for messages showing details within a method that, for example,
+   * returns a value. (So, for example, the NCM_Task method IsTaskWorthwhile() would primarily use
+   * finest, as it just returns a boolean, and the logs within it are just for details) Finer for
+   * just about everything else. (There's also the SERVER, INFO, etc. levels) Just keep these things
+   * in mind while adding new logging code.
    */
   public static void log(final Level level, final String message, final @Nullable Throwable t) {
     final ProLogSettings settings = ProLogSettings.loadSettings();
